% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_numbers_above_the_bars}
\alias{add_numbers_above_the_bars}
\title{Add numbers horizontally above the bars of a barplot}
\usage{
add_numbers_above_the_bars(stats, bp, color)
}
\arguments{
\item{stats}{a numeric vector}

\item{bp}{the result of \strong{\code{barplot}} command, usually a numeric
vector or matrix}

\item{color}{string. The color for the numbers}
}
\description{
Add numbers horizontally above the bars of a barplot
}
