% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{add_performance_biomarkers}
\alias{add_performance_biomarkers}
\title{adds one more row to the `biomarkers.synergy.res` data.frame with the
performance-related biomarkers}
\usage{
add_performance_biomarkers(biomarkers.synergy.res, biomarkers.active,
  biomarkers.inhibited)
}
\description{
adds one more row to the `biomarkers.synergy.res` data.frame with the
performance-related biomarkers
}
