% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{add_row_to_biomarkers_df}
\alias{add_row_to_biomarkers_df}
\title{Add row to a biomarkers \code{data.frame}}
\usage{
add_row_to_biomarkers_df(biomarkers.df, biomarkers.active,
  biomarkers.inhibited, row.name = "PERF")
}
\arguments{
\item{biomarkers.df}{a \code{data.frame} object with values only in the
the 3-element set \{-1,0,1\}. The column names should be node names
(gene, protein names, etc.).}

\item{biomarkers.active}{a character vector whose elements are nodes that
were found as active biomarkers. These biomarkers should be a subset of the
column names (nodes) of the \code{biomarkers.df}.}

\item{biomarkers.inhibited}{a character vector whose elements are nodes that
were found as inhibited biomarkers. These biomarkers should be a subset of the
column names (nodes) of the \code{biomarkers.df}.}

\item{row.name}{string. The name of the new row that we will added.
Default value: 'PERF' (referring to 'performance biomarkers')}
}
\description{
Use this function on a biomarkers \code{data.frame} object (with values only
in the 3-element set \{-1,0,1\}) to add an extra first row with values from
the \emph{active} and \emph{inhibited} biomarkers given (\code{biomarkers.active}
and \code{biomarkers.inhibited} parameters).
}
