% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{assign_link_operator_value_to_equation}
\alias{assign_link_operator_value_to_equation}
\title{Assign link operator value to boolean equation}
\usage{
assign_link_operator_value_to_equation(equation)
}
\arguments{
\item{equation}{string. The boolean equation in the form
\eqn{Target *= (Activator or Activator or...)
and not (Inhibitor or Inhibitor or...)"}}
}
\value{
\strong{1} if the \code{equation} has the '\emph{or not}' link operator,
\strong{0} if the \code{equation} has the '\emph{and not}' link operator and
\strong{NA} if it has neither.
}
\description{
Assign link operator value to boolean equation
}
