% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{biomarker_tp_analysis}
\alias{biomarker_tp_analysis}
\title{Biomarker analysis based on TP model classification}
\usage{
biomarker_tp_analysis(
  model.predictions,
  models.stable.state,
  models.link.operator = NULL,
  observed.synergies,
  threshold
)
}
\arguments{
\item{model.predictions}{a \code{data.frame} object with rows the models and
columns the drug combinations. Possible values for each \emph{model-drug combination
element} are either \emph{0} (no synergy predicted), \emph{1} (synergy was
predicted) or \emph{NA} (couldn't find stable states in either the drug
combination inhibited model or in any of the two single-drug inhibited models).}

\item{models.stable.state}{a matrix (nxm) with n models and m nodes. The row
names of the matrix specify the models' names whereas the column names
specify the name of the network nodes (gene, proteins, etc.).
Possible values for each \emph{model-node element}
are either \emph{0} (inactive node) or \emph{1} (active node). Note that the
rows (models) have to be in the same order as in the \code{model.predictions}
parameter.}

\item{models.link.operator}{a matrix (nxm) with n models and m nodes. The row
names of the matrix specify the models' names whereas the column names specify
the name of the network nodes (gene, proteins, etc.). Possible values for each
\emph{model-node element} are either \emph{0} (\strong{AND NOT} link operator),
\emph{1} (\strong{OR NOT} link operator) or \emph{0.5} if the node is not targeted
by both activating and inhibiting regulators (no link operator). Default value:
NULL (no analysis on the models parameterization regarding the mutation of the
boolean equation link operator will be done).}

\item{observed.synergies}{a character vector with elements the names of the
drug combinations that were found as synergistic. This should be a subset of
the tested drug combinations, that is the column names of the \code{model.predictions}
parameter.}

\item{threshold}{numeric. A number in the [0,1] interval, above which (or
below its negative value) a biomarker will be registered in the returned result.
Values closer to 1 translate to a more strict threshold and thus less
biomarkers are found.}
}
\value{
a list with various elements:
\itemize{
  \item \code{observed.model.predictions}: the part of the \code{model.predictions}
  data that includes the \code{observed.synergies}.
  \item \code{unobserved.model.predictions}: the complementary part of the
  \code{model.predictions} data that does not include the \code{observed.synergies}
  \item \code{predicted.synergies}: a character vector of the synergies (drug
  combination names) that were predicted by \strong{at least one} of the models
  in the dataset.
  \item \code{synergy.subset.stats}: an integer vector with elements the number
  of models the predicted each \strong{observed synergy subset}.
  \item \code{models.synergies.tp}: an integer vector of true positive (TP)
  values, one for each model.
  \item \code{diff.tp.mat}: a matrix whose rows are \strong{vectors of
  average node activity state differences} between two groups of models where
  the classification was based on the number of true positive predictions.
  Rows represent the different classification group matchings, e.g. (1,2) means the
  models that predicted 1 TP synergy vs the models that predicted 2 TP
  synergies and the columns represent the network's node names.
  Values are in the [-1,1] interval.
  \item \code{biomarkers.tp.active}: a character vector whose elements are
  the names of the \emph{active state} biomarkers. These nodes appear as more
  active in the better performance models.
  \item \code{biomarkers.tp.inhibited}: a character vector whose elements are
  the names of the \emph{inhibited state} biomarkers. These nodes appear as more
  inhibited in the better performance models.
  \item \code{diff.link.tp.mat}: a matrix whose rows are \strong{vectors of
  average node link operator differences} between two groups of models where
  the classification was based on the number of true positive predictions.
  Rows represent the different classification group matchings, e.g. (1,2) means the
  models that predicted 1 TP synergy vs the models that predicted 2 TP
  synergies and the columns represent the network's node names.
  Values are in the [-1,1] interval.
  \item \code{biomarkers.tp.or}: a character vector whose elements are
  the names of the \emph{OR} link operator biomarkers. These nodes have
  mostly the \emph{OR} link operator in their respective boolean equations
  in the better performance models.
  \item \code{biomarkers.tp.and}: a character vector whose elements are
  the names of the \emph{AND} link operator biomarkers. These nodes have
  mostly the \emph{AND} link operator in their respective boolean equations
  in the better performance models.
}
}
\description{
Use this function to perform a full biomarker analysis on an ensemble boolean model
dataset where the model classification is based on the number of \emph{true
positive} (TP) predictions. This analysis enables the discovery of \emph{performance
biomarkers}, nodes whose activity and/or boolean model parameterization (link
operator) affects the prediction performance of the models (as measured by
the number of TPs).
}
\seealso{
Other general analysis functions: 
\code{\link{biomarker_mcc_analysis}()},
\code{\link{biomarker_synergy_analysis}()}
}
\concept{general analysis functions}
