% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emba.R
\docType{package}
\name{emba}
\alias{emba}
\alias{emba-package}
\title{emba}
\description{
A collection of useful functions that help the analysis and discovery of
biomarkers in an ensemble boolean model dataset
}
\details{
For a complete list of functions, use \code{library(help = "emba")}
}
