% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_alt_drugname}
\alias{get_alt_drugname}
\title{Get alternative drug combination name}
\usage{
get_alt_drugname(drug.comb)
}
\arguments{
\item{drug.comb}{a string in the form \emph{drugname.1-drugname.2} (no
spaces between the names and the hyphen '-')}
}
\value{
the alternative, yet equivalent drug combination
}
\description{
Use this function on a string \emph{A-B} that represents a drug combination,
to get the reverse combination name - \emph{B-A} - for testing/checking data.
}
\examples{
drug.comb = "A-B"
alt.drug.comb = get_alt_drugname(drug.comb)

}
