% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_avg_activity_diff_based_on_diff_synergy_prediction}
\alias{get_avg_activity_diff_based_on_diff_synergy_prediction}
\title{To get meaningful results, one set must be a subset of the other
Example use:
synergy.set.str = "A-B,A-D,B-D,P-S"
synergy.subset.str = "A-B,B-D,P-S"}
\usage{
get_avg_activity_diff_based_on_diff_synergy_prediction(synergy.set.str,
  synergy.subset.str, model.predictions, models.stable.state)
}
\description{
To get meaningful results, one set must be a subset of the other
Example use:
synergy.set.str = "A-B,A-D,B-D,P-S"
synergy.subset.str = "A-B,B-D,P-S"
}
\seealso{
Other average data difference functions: \code{\link{get_avg_activity_diff_based_on_mcc_clustering}},
  \code{\link{get_avg_activity_diff_based_on_specific_synergy_prediction}},
  \code{\link{get_avg_activity_diff_based_on_tp_predictions}},
  \code{\link{get_avg_activity_diff_mat_based_on_mcc_clustering}},
  \code{\link{get_avg_activity_diff_mat_based_on_tp_predictions}},
  \code{\link{get_avg_link_operator_diff_mat_based_on_mcc_clustering}},
  \code{\link{get_avg_link_operator_diff_mat_based_on_tp_predictions}}
}
\concept{average data difference functions}
