% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_avg_activity_diff_based_on_mcc_classification}
\alias{get_avg_activity_diff_based_on_mcc_classification}
\title{TO BE DELETED !!! ???
`class.id.low` is the `mcc.interval` id for the 'bad' models
`class.id.high` is the `mcc.interval` id for the 'good' models}
\usage{
get_avg_activity_diff_based_on_mcc_classification(models, models.mcc,
  mcc.intervals, models.stable.state, class.id.low, class.id.high)
}
\description{
TO BE DELETED !!! ???
`class.id.low` is the `mcc.interval` id for the 'bad' models
`class.id.high` is the `mcc.interval` id for the 'good' models
}
