% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_avg_activity_diff_based_on_specific_synergy_prediction}
\alias{get_avg_activity_diff_based_on_specific_synergy_prediction}
\title{Example use: `drug.comb` = "AK-PD"}
\usage{

  get_avg_activity_diff_based_on_specific_synergy_prediction(synergy.set.str,
  synergy.subset.str, model.predictions, models.stable.state)


  get_avg_activity_diff_based_on_specific_synergy_prediction(synergy.set.str,
  synergy.subset.str, model.predictions, models.stable.state)
}
\description{
Example use: `drug.comb` = "AK-PD"

To get meaningful results, one set must be a subset of the other
Example use:
synergy.set.str = "A-B,A-D,B-D,P-S"
synergy.subset.str = "A-B,B-D,P-S"
}
