% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_avg_link_operator_diff_based_on_synergy_set_cmp}
\alias{get_avg_link_operator_diff_based_on_synergy_set_cmp}
\title{Get the average link operator difference based on the comparison of two synergy sets}
\usage{
get_avg_link_operator_diff_based_on_synergy_set_cmp(
  synergy.set.str,
  synergy.subset.str,
  model.predictions,
  models.link.operator,
  penalty = 0
)
}
\arguments{
\item{synergy.set.str}{a string of drug combinations, comma-separated. The
number of the specified combinations must be larger than the ones defined
in the \code{synergy.subset.str} parameter. They also must be included in the
tested drug combinations, i.e. the columns of the \code{model.predictions}
parameter.}

\item{synergy.subset.str}{a string of drug combinations, comma-separated.
There must be at least one combination defined and all of them should also
be included in the \code{synergy.set.str} parameter.}

\item{model.predictions}{a \code{data.frame} object with rows the models and
columns the drug combinations. Possible values for each \emph{model-drug combination
element} are either \emph{0} (no synergy predicted), \emph{1} (synergy was
predicted) or \emph{NA} (couldn't find stable states in either the drug
combination inhibited model or in any of the two single-drug inhibited models)}

\item{models.link.operator}{a \code{data.frame} (nxm) with n models and m nodes.
The row names specify the models' names whereas the column names specify the
network nodes (gene, proteins, etc.). Possible values for each \emph{model-node
element} are either \emph{0} (\strong{AND NOT} link operator), \emph{1}
(\strong{OR NOT} link operator) or \emph{0.5} if the node is not targeted by
both activating and inhibiting regulators (no link operator).}

\item{penalty}{value between 0 and 1 (inclusive). A value of 0 means no
penalty and a value of 1 is the strickest possible penalty. Default value is 0.
This penalty is used as part of a weighted term to the difference in a value of
interest (e.g. activity or link operator difference) between two group of
models, to account for the difference in the number of models from each
respective model group.}
}
\value{
a numeric vector with values in the [-1,1] interval (minimum and
maximum possible average difference) and with the names attribute
representing the name of the nodes.
}
\description{
This function uses the \code{\link{get_avg_activity_diff_based_on_synergy_set_cmp}}
which splits the models to 'good' and 'bad' based on the predictions
of two different synergy sets, one of them being a subset of the other.
The 'good' models are those that predict the \code{synergy.set.str}
(e.g. "A-B,A-C,B-C") while the 'bad' models are those that predict the
\code{synergy.subset.str} (e.g. "A-B,B-C"). Then, for each network node,
the function finds the node's average link operator value in each of the two classes
(a value in the [0,1] interval, 0 being \emph{AND NOT} and 1 being \emph{OR NOT})
and then subtracts the bad class average link operator value from the good one,
taking into account the given \code{penalty} factor and the number of models
in the 'good' and 'bad' class respectively.
}
\section{Details}{

So, if a node has a value close to -1 it means that on average,
this node's boolean equation has the \strong{AND NOT} link operator in the
models that predicted the extra synergy(-ies) that are included in the
\code{synergy.set.str} but not in the \code{synergy.subset.str}, whereas a
value closer to 1 means that the node's boolean equation has mostly the
\strong{OR NOT} link operator in these models.
These nodes are potential \strong{link operator biomarkers} because the
structure of their respective boolean equations (denoted by their link operator)
can influence the prediction performance of a model and make it predict the
extra synergy(-ies). A value closer to 0 indicates that the link operator in
the node's boolean equation is \strong{not so much different} between the
models that predicted the synergy set and those that predicted it's subset,
so it won't not be a node of interest when searching for potential link operator
biomarkers for the extra synergy(-ies).
A value exactly equal to 0 can also mean that this node didn't not have a link operator
in its boolean equation, again making it a non-important indicator of difference
in model performance.
}

\seealso{
\code{\link{get_vector_diff}}

Other average data difference functions: 
\code{\link{get_avg_activity_diff_based_on_mcc_clustering}()},
\code{\link{get_avg_activity_diff_based_on_specific_synergy_prediction}()},
\code{\link{get_avg_activity_diff_based_on_synergy_set_cmp}()},
\code{\link{get_avg_activity_diff_based_on_tp_predictions}()},
\code{\link{get_avg_activity_diff_mat_based_on_mcc_clustering}()},
\code{\link{get_avg_activity_diff_mat_based_on_specific_synergy_prediction}()},
\code{\link{get_avg_activity_diff_mat_based_on_tp_predictions}()},
\code{\link{get_avg_link_operator_diff_mat_based_on_mcc_clustering}()},
\code{\link{get_avg_link_operator_diff_mat_based_on_specific_synergy_prediction}()},
\code{\link{get_avg_link_operator_diff_mat_based_on_tp_predictions}()}
}
\concept{average data difference functions}
