% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_biomarkers_per_synergy}
\alias{get_biomarkers_per_synergy}
\title{returns a data.frame, where the columns represent the network nodes, the rows
represent the predicted synergies and the values can either 1 (active biomarker),
-1 (inhibited biomarker) or 0 (not a biomarker)}
\usage{
get_biomarkers_per_synergy(predicted.synergies, biomarkers.dir, models.dir)
}
\description{
returns a data.frame, where the columns represent the network nodes, the rows
represent the predicted synergies and the values can either 1 (active biomarker),
-1 (inhibited biomarker) or 0 (not a biomarker)
}
