% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_edges_from_topology_file}
\alias{get_edges_from_topology_file}
\title{Get the edges from a specified topology}
\usage{
get_edges_from_topology_file(topology.file)
}
\arguments{
\item{topology.file}{string. The name of the .sif file (can be a full path
name).}
}
\value{
a matrix with as many rows as in the .sif topology file (each row is
an edge) and 4 columns defining the source and target node name, the
regulation (activation or inhibition) and the color (green or red) of the
signed interaction.
}
\description{
Use this function to read a topology .sif file (either space or tab-delimited)
and get a matrix of network edges specifying the source and target name, the
regulation effect (activation or inhibition) and the color (green or red) of
each interaction.
}
\examples{
topology.file = system.file("extdata", "example.sif", package = "emba", mustWork = TRUE)
edges = get_edges_from_topology_file(topology.file)

}
