% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_model_names}
\alias{get_model_names}
\title{Get the model names}
\usage{
get_model_names(models.dir)
}
\arguments{
\item{models.dir}{string. A dir with \emph{.gitsbe} files/models}
}
\value{
a character vector of the model names, corresponding to the names
of the \emph{.gitsbe} files.
}
\description{
Get the model names
}
\examples{

models.dir = system.file("extdata", "models", package = "emba", mustWork = TRUE)
models = get_model_names(models.dir)

}
