% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_models_based_on_mcc_interval}
\alias{get_models_based_on_mcc_interval}
\title{TO BE DELETED !!! ???}
\usage{
get_models_based_on_mcc_interval(models, models.mcc, mcc.interval,
  include.high.value = FALSE)
}
\description{
TO BE DELETED !!! ???
}
