% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get neighbor nodes}
\usage{
get_neighbors(net, nodes)
}
\arguments{
\item{net}{igraph object}

\item{nodes}{character vector of node names}
}
\value{
a character vector of all the unique neighbors of the given
\code{nodes} in the \code{net} graph.
}
\description{
Given an igraph network object and vector of node names, this function
returns the set of unique neighbor nodes considering both ingoing and outgoing
edges (the closed neighbourhood node set).
}
