% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_node_colors}
\alias{get_node_colors}
\title{`net` is an igraph network object with node labels as: `V(net)$name`}
\usage{
get_node_colors(net, diff, col)
}
\description{
`net` is an igraph network object with node labels as: `V(net)$name`
}
