% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_node_names}
\alias{get_node_names}
\title{Get the node names}
\usage{
get_node_names(models.dir)
}
\arguments{
\item{models.dir}{string. A directory with at least one \emph{.gitsbe} file/model.}
}
\value{
a character vector of the node names (protein and/or gene names)
}
\description{
This function uses the first \emph{.gitsbe} file that it finds inside the given
directory to output a vector of the network node names (which should be the
same for every model)
}
\examples{

models.dir = system.file("extdata", "models", package = "emba", mustWork = TRUE)
nodes = get_node_names(models.dir)

}
