% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_perf_biomarkers_per_cell_line}
\alias{get_perf_biomarkers_per_cell_line}
\title{`biomarkers.dirs` is a vector of the cell lines' biomarker directories
and `type` can be either 'active' or 'inhibited'}
\usage{
get_perf_biomarkers_per_cell_line(biomarkers.dirs, type)
}
\description{
`biomarkers.dirs` is a vector of the cell lines' biomarker directories
and `type` can be either 'active' or 'inhibited'
}
