% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_synergy_scores}
\alias{get_synergy_scores}
\title{Get synergy scores from a file}
\usage{
get_synergy_scores(file_name)
}
\arguments{
\item{file_name}{string. The name of the file, can be a full path.}
}
\value{
a \code{data.frame} containing a representation of the data in the file.
}
\description{
Use this function to read a file that has synergy scoring data for a list
of drug perturbations. The first line of the file is always skipped (no matter
if you have a header or data in it) and it's basic structure must be two columns,
the first being the names of drug combinations and the second the synergy
scores for those perturbations. The columns must be \emph{tab-separated}.
See example below for the format of such a file generated by the
\href{https://druglogics.github.io/druglogics-doc/drabme.html}{Drabme} module.
}
\examples{

ensemble_wise_synergies_file = system.file("extdata", "ensemble_wise_synergies",
  package = "emba", mustWork = TRUE)
synergy_scores = get_synergy_scores(ensemble_wise_synergies_file)

}
