% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_x_axis_values}
\alias{get_x_axis_values}
\title{Get the refined x-axis values}
\usage{
get_x_axis_values(models.stats, there.is.one.NaN.category, cont.values)
}
\arguments{
\item{models.stats}{table object, the result of using \link[base]{table} on
a (numeric) vector. Usually it represents some models statistics summary -
counts for each TP prediction value for example.}

\item{there.is.one.NaN.category}{logical. Is there one \emph{NaN} category?
(check is done before on the \emph{names} attribute of the \code{models.stats})}

\item{cont.values}{logical. If TRUE, the values of the x-axis will be trimmed
to 3 digits after the decimal point. Otherwise, they will be returned as they
are.}
}
\description{
This function returns the x-axis values that are going to be used by
\link[emba]{make_barplot_on_models_stats} to render the bar plot.
}
