% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_avg_state_diff_graphs}
\alias{plot_avg_state_diff_graphs}
\title{Plot the graphs of many average state differences vectors}
\usage{
plot_avg_state_diff_graphs(net, diff.df, layout)
}
\arguments{
\item{net}{igraph graph object}

\item{diff.df}{a data.frame whose rows are vectors of average node activity state
differences between two groups of models based on some kind of classification
(e.g. number of TP predictions) and whose names are set in the \code{rownames}
attribute of the data frame. The columns represent the network's node names.}

\item{layout}{a (nx2) numeric matrix of x-y coordinates (2 columns) for each
of the nodes (n) in the \code{net} igraph object}
}
\description{
This function presents a convenient way to use many times the
\code{\link{plot_avg_state_diff_graph}} function.
}
\seealso{
Other network ploting functions: \code{\link{plot_avg_state_diff_graph_vis}},
  \code{\link{plot_avg_state_diff_graph}}
}
\concept{network ploting functions}
