% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network_vis}
\alias{plot_network_vis}
\title{plot network using the `visNetwork` library}
\usage{
plot_network_vis(net, diff, layout, title)
}
\arguments{
\item{net}{a net}

\item{diff}{a diff}

\item{layout}{layout}

\item{title}{a title}
}
\description{
A description
}
