% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print_model_and_drug_stats}
\alias{print_model_and_drug_stats}
\title{Print model and drug statistics}
\usage{
print_model_and_drug_stats(drug.combs, models, nodes, html.output)
}
\arguments{
\item{drug.combs}{integer. Number of drug combinations tested}

\item{models}{integer. Number of models tested}

\item{nodes}{integer. Number of network nodes}

\item{html.output}{logical. If TRUE, the printed output will look nice in an
HTML document}
}
\description{
Use this function to pretty print in an R notebook useful statistics for the
ensemble model analysis: how many drug combinations were tested by each model,
the number of models used and how many nodes each boolean network model had.
}
