% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{update_biomarker_files}
\alias{update_biomarker_files}
\title{We get the (previously-found) biomarkers from the respective file.
There are 3 policies regarding what to do with the 'new' biomarkers
when they share common nodes with the 'old' biomarkers. These are
given by the `method`:
1) `replace` (default): if there is at least one new biomarker
common with the previously found ones, we discard the old ones and
write to the file the new biomarkers
2) `prune.to.common`: keep only the common biomarkers
3) `extend`: we add the non-common biomarkers to the old ones}
\usage{
update_biomarker_files(biomarkers.dir, drug.comb, biomarkers.active.new,
  biomarkers.inhibited.new, method = "replace")
}
\description{
Note: No matter the `method`, if there are no common biomarkers
between old and new, we just add the new ones
}
