% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{validate_observed_synergies_data}
\alias{validate_observed_synergies_data}
\title{Validate observed synergies data}
\usage{
validate_observed_synergies_data(observed.synergies,
  drug.combinations.tested)
}
\arguments{
\item{observed.synergies}{a character vector of drug combinations}

\item{drug.combinations.tested}{a character vector of drug combinations}
}
\value{
NULL if no errors found, otherwise stops execution.
}
\description{
This function checks that the observed synergies are part (a subset) of the
tested drug combinations
}
