% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emba.R
\docType{package}
\name{emba}
\alias{emba}
\alias{emba-package}
\title{emba}
\description{
Analysis and visualization of an ensemble of
boolean models for biomarker discovery in cancer cell networks.
}
\details{
For a complete list of functions, use \code{library(help = "emba")}
}
