% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_avg_activity_diff_based_on_mcc_classification}
\alias{get_avg_activity_diff_based_on_mcc_classification}
\title{`class.id.low` is the `mcc.interval` id for the 'bad' models
`class.id.high` is the `mcc.interval` id for the 'good' models}
\usage{
get_avg_activity_diff_based_on_mcc_classification(models, models.mcc,
  mcc.intervals, models.stable.state, class.id.low, class.id.high)
}
\description{
`class.id.low` is the `mcc.interval` id for the 'bad' models
`class.id.high` is the `mcc.interval` id for the 'good' models
}
\seealso{
Other get average activity difference functions: \code{\link{get_avg_activity_diff_based_on_diff_synergy_prediction}},
  \code{\link{get_avg_activity_diff_based_on_mcc_clustering}},
  \code{\link{get_avg_activity_diff_based_on_specific_synergy_prediction}},
  \code{\link{get_avg_activity_diff_based_on_tp_predictions}}
}
\concept{get average activity difference functions}
