% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_avg_activity_diff_based_on_mcc_clustering}
\alias{get_avg_activity_diff_based_on_mcc_clustering}
\title{Title mcc}
\usage{
get_avg_activity_diff_based_on_mcc_clustering(models.mcc,
  models.mcc.no.nan.sorted, models.stable.state, mcc.class.ids,
  models.cluster.ids, class.id.low, class.id.high)
}
\description{
Title mcc
}
\seealso{
Other average activity difference functions: \code{\link{get_avg_activity_diff_based_on_diff_synergy_prediction}},
  \code{\link{get_avg_activity_diff_based_on_mcc_classification}},
  \code{\link{get_avg_activity_diff_based_on_specific_synergy_prediction}}
}
\concept{average activity difference functions}
