% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_avg_activity_diff_based_on_specific_synergy_prediction}
\alias{get_avg_activity_diff_based_on_specific_synergy_prediction}
\title{Example use: `drug.comb` = "AK-PD"}
\usage{

  get_avg_activity_diff_based_on_specific_synergy_prediction(model.predictions,
  models.stable.state, drug.comb)
}
\description{
Example use: `drug.comb` = "AK-PD"
}
\seealso{
Other average data difference functions: \code{\link{get_avg_activity_diff_based_on_diff_synergy_prediction}},
  \code{\link{get_avg_activity_diff_based_on_mcc_clustering}},
  \code{\link{get_avg_activity_diff_based_on_tp_predictions}},
  \code{\link{get_avg_activity_diff_mat_based_on_mcc_clustering}},
  \code{\link{get_avg_activity_diff_mat_based_on_tp_predictions}},
  \code{\link{get_avg_link_operator_diff_mat_based_on_mcc_clustering}},
  \code{\link{get_avg_link_operator_diff_mat_based_on_tp_predictions}}
}
\concept{average data difference functions}
