% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_avg_activity_diff_based_on_tp_predictions}
\alias{get_avg_activity_diff_based_on_tp_predictions}
\title{Get the average activity difference based on the number of true positives}
\usage{
get_avg_activity_diff_based_on_tp_predictions(models, models.synergies.tp,
  models.stable.state, num.low, num.high)
}
\arguments{
\item{models}{s}

\item{models.synergies.tp}{a}

\item{models.stable.state}{ert}

\item{num.low}{integer. The number of true positives representing the 'bad'
model class}

\item{num.high}{integer. The number of true positives representing the 'good'
model class. This number has to be strictly higher than \code{num.low}}
}
\value{

}
\description{
This function splits the models to good and bad based on the number of true
positive predictions: \emph{num.high} TPs (good) vs \emph{num.low} TPs (bad).
Then, for each network node, finds the average activity in each of the two
classes and then subtracts the 'bad' average activity value from the good
}
