% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{get_avg_activity_diff_mat_based_on_tp_predictions}
\alias{get_avg_activity_diff_mat_based_on_tp_predictions}
\title{Get average activity difference matrix based on the number of true positives}
\usage{
get_avg_activity_diff_mat_based_on_tp_predictions(models,
  models.synergies.tp, models.stable.state)
}
\arguments{
\item{models}{character vector. The model names.}

\item{models.synergies.tp}{an integer vector of TP values. The \emph{names}
attribute holds the models' names and have to be in the same order as in the
\code{models} parameter.}

\item{models.stable.state}{a matrix (nxm) with n models and m nodes. The row
names of the matrix specify the models' names (same order as in the \code{models}
parameter) whereas the column names specify the name of the network nodes
(gene, proteins, etc.). Possible values for each \emph{model-node element}
are either \emph{0} (inactive node) or \emph{1} (active node).}
}
\value{
a matrix whose rows are \strong{vectors of average node activity
state differences} between two groups of models based on some kind of classification
(e.g. number of TP predictions) and whose names are set in the \code{rownames}
attribute of the matrix (usually denoting the diffferent classification
groups, e.g. (1,2) means the models that predicted 1 TP synergy vs the models
that predicted 2 TP synergies, if the classification is done by number of TP
predictions). The columns represent the network's node names. Values are in
the [-1,1] interval.
}
\description{
This function uses the \code{\link[emba]{get_avg_activity_diff_based_on_tp_predictions}}
function on each row vector of the given \code{models.stable.state} matrix.
}
\seealso{
Other average data difference functions: \code{\link{get_avg_activity_diff_based_on_tp_predictions}},
  \code{\link{get_avg_link_operator_diff_mat_based_on_tp_predictions}}
}
\concept{average data difference functions}
