% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_synergy_comparison_sets}
\alias{get_synergy_comparison_sets}
\title{helper function to check which synergy sets to compare (the small synergy set
misses just one synergy from the larger set)}
\usage{
get_synergy_comparison_sets(synergy.subset.stats)
}
\description{
helper function to check which synergy sets to compare (the small synergy set
misses just one synergy from the larger set)
}
