% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_barplot_on_models_stats}
\alias{make_barplot_on_models_stats}
\title{models.stats is the result of `table(num.vector)`
Use `threshold` when there too many categories and the graph appears too dense
`cont.values` is used for trimming the digits of continuous values on the x-axis}
\usage{
make_barplot_on_models_stats(models.stats, cell.line, title, xlab, ylab,
  cont.values = FALSE, threshold = 0)
}
\description{
models.stats is the result of `table(num.vector)`
Use `threshold` when there too many categories and the graph appears too dense
`cont.values` is used for trimming the digits of continuous values on the x-axis
}
