% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mcc_classes_hist}
\alias{plot_mcc_classes_hist}
\title{Plot histogram of the MCC classes}
\usage{
plot_mcc_classes_hist(models.mcc.no.nan.sorted, models.cluster.ids,
  num.of.classes, mcc.class.ids)
}
\description{
Plot histogram of the MCC classes
}
