% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print_biomarkers_per_predicted_synergy}
\alias{print_biomarkers_per_predicted_synergy}
\title{yeah baby}
\usage{
print_biomarkers_per_predicted_synergy(biomarkers.dir, drug.comb,
  predicted.synergies, html.output = TRUE)
}
\description{
yeah baby
}
